inventory-allocator/src/requirements.txt
click
pyyaml
pytest
pytest-cov
hypothesis
